feat: make runtime never import typescript
