+++ src/lib/layout/layout.ts
@@ -3,7 +3,7 @@ import { stripIndent } from 'common-tags'
 import * as FS from 'fs-jetpack'
 import * as Path from 'path'
 import { PackageJson } from 'type-fest'
-import { ParsedCommandLine } from 'typescript'
+import type { ParsedCommandLine } from 'typescript'
 import { findFile, findFileRecurisvelyUpwardSync } from '../../lib/fs'
 import { START_MODULE_NAME } from '../../runtime/start/start-module'
 import { rootLogger } from '../nexus-logger'
