+++ src/lib/plugin/utils.ts
@@ -1,12 +1,7 @@
 import { stripIndent } from 'common-tags'
 import { PackageJson } from 'type-fest'
-import * as Layout from '../layout'
-import { rootLogger } from '../nexus-logger'
 import { fatal } from '../process'
 import { Manifest, Plugin } from './types'
-import * as Reflection from '../reflection/reflect'
-
-const log = rootLogger.child('plugin')
 
 /**
  * Normalize a raw plugin manifest.
@@ -41,31 +36,5 @@ export function entrypointToManifest(plugin: Plugin): Manifest {
     `,
       { plugin }
     )
-  }
-}
-
-/**
- * This gets all the plugins in use in the app.
- *
- * @remarks
- *
- * This is useful for the CLI to get worktime plugins. This will run the app in
- * data mode, in this process.
- */
-export async function getUsedPlugins(layout: Layout.Layout): Promise<Plugin[]> {
-  try {
-    const reflection = await Reflection.reflect(layout, { usedPlugins: true, onMainThread: true })
-
-    if (!reflection.success) {
-      throw reflection.error
-    }
-
-    log.trace('got used plugins', { validPlugins: reflection.plugins })
-
-    return reflection.plugins
-  } catch (e) {
-    fatal('Failed to scan app for used plugins because there is a runtime error in the app', {
-      error: e,
-    })
   }
 }
