+++ src/runtime/app.ts
@@ -2,7 +2,7 @@ import * as Logger from '@nexus/logger'
 import type { MissingType, NexusGraphQLSchema } from '@nexus/schema/dist/core'
 import * as Plugin from '../lib/plugin'
 import { RuntimeContributions } from '../lib/plugin'
-import * as Reflection from '../lib/reflection'
+import * as Reflection from '../lib/reflection/stage'
 import { Index } from '../lib/utils'
 import * as Schema from './schema'
 import * as Server from './server'
